requirements.txt
ninja
trimesh
tensorboardX
torch
numpy 
matplotlib
PyMCubes
rich
dearpygui
packaging
scipy
lpips
imageio
torchmetrics
ruamel.yaml
pyrender